refactor: split validation logic into dedicated subpackage with modular files
